Permission handle improved and code refactored

+++ app/src/main/java/com/robin729/aqi/model/weather/Main.kt
@@ -3,4 +3,4 @@ package com.robin729.aqi.model.weather
 import com.google.gson.annotations.SerializedName
 
 class Main (@SerializedName("temp")
-           val temp: Int)
+           val temp: Float)
